+++ CHANGES.md
@@ -6,7 +6,7 @@
 - [TODO] Fix `strfmt` to support the default width qualifier for `%f`
 
 
-## 0.3.0 / 2016-October-??
+## 0.3.0 / 2016-October-13
 
 - Fix #1 `strfmt` to accept format-string as non-literal, eval'ed as a string at compile time
 - Efficiently repeat a string: `repstr`